Harden parent workflow tool installs
Restrict parent CI workflow token permissions, pin markdownlint-cli2 and mbake,
and verify the downloaded act archive before installing it in act-validation.

Split the large tooling contract helper into focused domain modules while
keeping the public orchestration imports stable. Fix the Rust audit Makefile
contract so the raw Makefile assertion checks $(CARGO) audit and the parsed
recipe assertion matches make-parser output.
