+++ .github/workflows/act-validation.yml
@@ -5,11 +5,16 @@ on:
     branches: [main]
   pull_request:
 
+permissions:
+  contents: read
+
 jobs:
   act-validation:
     runs-on: ubuntu-latest
     env:
       ACT_VERSION: v0.2.80
+      MARKDOWNLINT_CLI2_VERSION: 0.22.1
+      MBAKE_VERSION: 1.4.6
     steps:
       - name: Check out repository
         uses: actions/checkout@08eba0b27e820071cde6df949e0beb9ba4906955 # v4
@@ -31,10 +36,17 @@ jobs:
 
       - name: Install template test tools
         run: |
-          curl -fsSL "https://github.com/nektos/act/releases/download/${ACT_VERSION}/act_Linux_x86_64.tar.gz" |
-            sudo tar -xz -C /usr/local/bin act
-          npm install -g markdownlint-cli2
-          uv tool install mbake
+          act_archive=act_Linux_x86_64.tar.gz
+          act_url="https://github.com/nektos/act/releases/download/${ACT_VERSION}/${act_archive}"
+          curl -fsSLo "${act_archive}" "${act_url}"
+          act_sha256="$(
+            curl -fsSL "https://api.github.com/repos/nektos/act/releases/tags/${ACT_VERSION}" |
+              python -c 'import json, sys; release = json.load(sys.stdin); print(next(asset["digest"].split(":", 1)[1] for asset in release["assets"] if asset["name"] == "act_Linux_x86_64.tar.gz"))'
+          )"
+          printf '%s  %s\n' "${act_sha256}" "${act_archive}" | sha256sum -c -
+          sudo tar -xzf "${act_archive}" -C /usr/local/bin act
+          npm install -g "markdownlint-cli2@${MARKDOWNLINT_CLI2_VERSION}"
+          uv tool install "mbake==${MBAKE_VERSION}"
 
       - name: Check Docker runtime
         run: docker info
