+++ tests/helpers/ci_contracts.py
@@ -1,0 +1,210 @@
+"""Validate rendered CI and act-validation workflow contracts.
+
+This module owns generated CI workflow assertions, including coverage-action
+inputs, act-validation workflow structure, and shared checkout credential
+checks.  Release workflow helpers import ``_checkout_steps_disable_credentials``
+from here so checkout security behaviour is asserted consistently.
+"""
+
+from __future__ import annotations
+
+from typing import Any
+
+from tests.helpers.generated_files import (
+    parse_yaml_mapping,
+    require_mapping,
+    require_sequence,
+)
+
+
+def assert_ci_coverage_action_contract(
+    *, ci_workflow: str, package_name: str, use_rust: bool
+) -> None:
+    """Assert generated CI coverage inputs used by act validation.
+
+    Parameters
+    ----------
+    ci_workflow : str
+        UTF-8 text of the generated CI workflow.
+    package_name : str
+        Generated Python import package name used to derive the coverage
+        artefact suffix.
+    use_rust : bool
+        Whether the rendered variant includes the optional Rust extension.
+
+    Returns
+    -------
+    None
+        The helper returns after the coverage action contract matches
+        expectations.
+
+    Raises
+    ------
+    AssertionError
+        Raised when checkout credentials, coverage action pinning, coverage
+        output settings, artefact naming, or Rust manifest inputs are wrong.
+
+    Examples
+    --------
+    Validate a rendered CI workflow's coverage step::
+
+        assert_ci_coverage_action_contract(
+            ci_workflow=ci_workflow,
+            package_name="example_pkg",
+            use_rust=True,
+        )
+    """
+    parsed_ci_workflow = _parse_ci_workflow(ci_workflow)
+    jobs = require_mapping(parsed_ci_workflow, "jobs", "CI workflow")
+    lint_test = require_mapping(jobs, "lint-test", "CI workflow jobs")
+    steps = require_sequence(lint_test, "steps", "CI lint-test job")
+    assert _checkout_steps_disable_credentials(steps), (
+        "expected CI checkout steps to disable credential persistence"
+    )
+    coverage_steps = [
+        step
+        for step in steps
+        if isinstance(step, dict) and step.get("name") == "Test and Measure Coverage"
+    ]
+    assert len(coverage_steps) == 1, "expected one shared coverage action step"
+    coverage_step = coverage_steps[0]
+    assert (
+        coverage_step.get("uses")
+        == "leynos/shared-actions/.github/actions/generate-coverage"
+        "@d400b079fb6a8fa92f7e7b6c57f3d1c92a4b2d54"
+    ), "expected CI to use the pinned shared coverage action"
+    coverage_inputs = require_mapping(coverage_step, "with", "coverage step")
+    assert coverage_inputs.get("output-path") == "coverage.xml", (
+        "expected CI coverage output path to match the act assertion"
+    )
+    assert coverage_inputs.get("format") == "cobertura", (
+        "expected CI coverage format to match the CodeScene upload"
+    )
+    assert coverage_inputs.get("artefact-name-suffix") == package_name.replace(
+        "_", "-"
+    ), "expected package-specific coverage artefact name suffix"
+    if use_rust:
+        assert coverage_inputs.get("cargo-manifest") == "rust_extension/Cargo.toml", (
+            "expected Rust variant to pass the extension manifest to coverage"
+        )
+    else:
+        assert "cargo-manifest" not in coverage_inputs, (
+            "expected pure-Python variant to omit Rust coverage inputs"
+        )
+
+
+def _parse_ci_workflow(ci_workflow: str) -> dict[str, Any]:
+    """Parse a generated CI workflow as a YAML mapping."""
+    return parse_yaml_mapping(ci_workflow, "CI workflow")
+
+
+def _assert_ci_workflow_contracts(
+    *, parsed_ci_workflow: dict[str, Any], ci_workflow: str, use_rust: bool
+) -> None:
+    """Assert generated CI workflow contracts.
+
+    Parameters
+    ----------
+    parsed_ci_workflow : dict[str, Any]
+        Parsed generated CI workflow mapping.
+    ci_workflow : str
+        UTF-8 text of the generated CI workflow, used for string-level contract
+        assertions.
+    use_rust : bool
+        Whether the rendered variant includes the optional Rust extension.
+    """
+    jobs = require_mapping(parsed_ci_workflow, "jobs", "CI workflow")
+    lint_test = require_mapping(jobs, "lint-test", "CI workflow jobs")
+    steps = require_sequence(lint_test, "steps", "CI lint-test job")
+    assert _checkout_steps_disable_credentials(steps), (
+        "expected generated CI workflow to disable checkout credentials"
+    )
+    assert "make check-fmt" in ci_workflow, (
+        "expected generated CI workflow to run the formatting gate"
+    )
+    assert "make lint" in ci_workflow, (
+        "expected generated CI workflow to run the lint gate"
+    )
+    assert "make typecheck" in ci_workflow, (
+        "expected generated CI workflow to run the typecheck gate"
+    )
+    assert "make audit" in ci_workflow, (
+        "expected generated CI workflow to run the dependency audit gate"
+    )
+    assert "make test WITH_ACT=1" not in ci_workflow, (
+        "expected generated main CI workflow to leave act validation to a "
+        "separate workflow"
+    )
+    assert "make build" in ci_workflow, (
+        "expected generated CI workflow to build the project before checks"
+    )
+    assert "leynos/shared-actions/.github/actions/generate-coverage" in ci_workflow, (
+        "expected generated CI workflow to use the shared coverage action"
+    )
+    if use_rust:
+        assert "leynos/shared-actions/.github/actions/setup-rust" in ci_workflow, (
+            "expected Rust variant CI to set up Rust"
+        )
+        assert "Cache Rust lint and test tools" in ci_workflow, (
+            "expected Rust variant CI to cache Rust tools"
+        )
+        assert "cargo-manifest: rust_extension/Cargo.toml" in ci_workflow, (
+            "expected Rust variant CI to pass the Rust manifest to coverage"
+        )
+        assert "cargo install --locked cargo-audit" in ci_workflow, (
+            "expected Rust variant CI to install cargo-audit"
+        )
+    else:
+        assert "setup-rust" not in ci_workflow, (
+            "expected pure-Python CI to omit Rust setup"
+        )
+        assert "cargo-manifest" not in ci_workflow, (
+            "expected pure-Python CI to omit Rust coverage inputs"
+        )
+        assert "cargo-audit" not in ci_workflow, (
+            "expected pure-Python CI to omit Rust audit installation"
+        )
+
+
+def _assert_act_validation_workflow_contracts(
+    *, act_validation_workflow: str, use_rust: bool
+) -> None:
+    """Assert generated act-validation workflow contracts."""
+    assert "name: Act Validation" in act_validation_workflow, (
+        "expected generated act-validation workflow to be named"
+    )
+    assert "ACT_VERSION: v0.2.80" in act_validation_workflow, (
+        "expected generated act-validation workflow to pin act"
+    )
+    assert "act_Linux_x86_64.tar.gz" in act_validation_workflow, (
+        "expected generated act-validation workflow to install act"
+    )
+    assert "docker info" in act_validation_workflow, (
+        "expected generated act-validation workflow to verify Docker"
+    )
+    assert "make test WITH_ACT=1" in act_validation_workflow, (
+        "expected generated act-validation workflow to run act-enabled tests"
+    )
+    if use_rust:
+        assert "leynos/shared-actions/.github/actions/setup-rust" in (
+            act_validation_workflow
+        ), "expected Rust variant act-validation workflow to set up Rust"
+    else:
+        assert "setup-rust" not in act_validation_workflow, (
+            "expected pure-Python act-validation workflow to omit Rust setup"
+        )
+
+
+def _checkout_steps_disable_credentials(steps: list[Any]) -> bool:
+    """Return whether all checkout steps disable credential persistence."""
+    checkout_steps = [
+        step
+        for step in steps
+        if isinstance(step, dict)
+        and str(step.get("uses", "")).startswith("actions/checkout@")
+    ]
+    return bool(checkout_steps) and all(
+        isinstance(step.get("with"), dict)
+        and step["with"].get("persist-credentials") is False
+        for step in checkout_steps
+    )
