+++ tests/test_parent_ci.py
@@ -29,6 +29,19 @@ def test_parent_ci_splits_application_and_act_validation_tests() -> None:
         REPO_ROOT / ".github" / "workflows" / "act-validation.yml"
     )
 
+    assert "permissions:\n  contents: read" in ci_workflow, (
+        "expected parent CI to restrict GITHUB_TOKEN to repository contents reads"
+    )
+    assert "MARKDOWNLINT_CLI2_VERSION: 0.22.1" in ci_workflow, (
+        "expected parent CI to pin markdownlint-cli2"
+    )
+    assert "MBAKE_VERSION: 1.4.6" in ci_workflow, "expected parent CI to pin mbake"
+    assert 'npm install -g "markdownlint-cli2@${MARKDOWNLINT_CLI2_VERSION}"' in (
+        ci_workflow
+    ), "expected parent CI to install pinned markdownlint-cli2"
+    assert 'uv tool install "mbake==${MBAKE_VERSION}"' in ci_workflow, (
+        "expected parent CI to install pinned mbake"
+    )
     assert "make test\n" in ci_workflow, (
         "expected parent CI to run the normal parent test gate"
     )
@@ -41,11 +54,23 @@ def test_parent_ci_splits_application_and_act_validation_tests() -> None:
     assert "make test WITH_ACT=1" not in ci_workflow, (
         "expected parent CI to leave act validation to a separate workflow"
     )
-    assert "ACT_VERSION: v0.2.80" in act_workflow, (
-        "expected parent act-validation workflow to pin the act release"
+    assert "permissions:\n  contents: read" in act_workflow, (
+        "expected parent act-validation workflow to restrict GITHUB_TOKEN"
     )
-    assert "act_Linux_x86_64.tar.gz" in act_workflow, (
-        "expected parent act-validation workflow to install the act Linux binary"
+    assert "ACT_VERSION:" in act_workflow, (
+        "expected parent act-validation workflow to declare an act version"
+    )
+    assert (
+        "act_Linux_x86_64.tar.gz" in act_workflow and "${ACT_VERSION}" in act_workflow
+    ), "expected parent act-validation workflow to build the act URL from ACT_VERSION"
+    assert "sha256sum -c -" in act_workflow, (
+        "expected parent act-validation workflow to verify the act archive checksum"
+    )
+    assert 'npm install -g "markdownlint-cli2@${MARKDOWNLINT_CLI2_VERSION}"' in (
+        act_workflow
+    ), "expected parent act-validation workflow to install pinned markdownlint-cli2"
+    assert 'uv tool install "mbake==${MBAKE_VERSION}"' in act_workflow, (
+        "expected parent act-validation workflow to install pinned mbake"
     )
     assert "uv tool install mdformat-all" not in act_workflow, (
         "expected parent act-validation workflow not to install mdformat-all through uv"
